app/requirements.txt
aiohappyeyeballs==2.6.1 ; python_full_version == "3.11.2"
aiohttp==3.12.14 ; python_full_version == "3.11.2"
aiosignal==1.4.0 ; python_full_version == "3.11.2"
annotated-types==0.7.0 ; python_full_version == "3.11.2"
apispec==6.8.2 ; python_full_version == "3.11.2"
async-timeout==5.0.1 ; python_full_version == "3.11.2"
asyncio==3.4.3 ; python_full_version == "3.11.2"
attrs==25.3.0 ; python_full_version == "3.11.2"
blinker==1.9.0 ; python_full_version == "3.11.2"
certifi==2025.7.14 ; python_full_version == "3.11.2"
charset-normalizer==3.4.2 ; python_full_version == "3.11.2"
click==8.2.1 ; python_full_version == "3.11.2"
colorama==0.4.6 ; python_full_version == "3.11.2" and platform_system == "Windows"
flask-smorest==0.46.1 ; python_full_version == "3.11.2"
flask==3.1.1 ; python_full_version == "3.11.2"
frozenlist==1.7.0 ; python_full_version == "3.11.2"
idna==3.10 ; python_full_version == "3.11.2"
itsdangerous==2.2.0 ; python_full_version == "3.11.2"
jinja2==3.1.6 ; python_full_version == "3.11.2"
load-dotenv==0.1.0 ; python_full_version == "3.11.2"
markupsafe==3.0.2 ; python_full_version == "3.11.2"
marshmallow==4.0.0 ; python_full_version == "3.11.2"
multidict==6.6.3 ; python_full_version == "3.11.2"
packaging==25.0 ; python_full_version == "3.11.2"
propcache==0.3.2 ; python_full_version == "3.11.2"
pydantic-core==2.33.2 ; python_full_version == "3.11.2"
pydantic==2.11.7 ; python_full_version == "3.11.2"
python-dotenv==1.1.1 ; python_full_version == "3.11.2"
redis==6.2.0 ; python_full_version == "3.11.2"
requests==2.32.4 ; python_full_version == "3.11.2"
setuptools==80.9.0 ; python_full_version == "3.11.2"
simplejson==3.20.1 ; python_full_version == "3.11.2"
tenacity==9.1.2 ; python_full_version == "3.11.2"
typing-extensions==4.14.1 ; python_full_version == "3.11.2"
typing-inspection==0.4.1 ; python_full_version == "3.11.2"
urllib3==2.5.0 ; python_full_version == "3.11.2"
uwsgi==2.0.30 ; python_full_version == "3.11.2"
webargs==8.7.0 ; python_full_version == "3.11.2"
werkzeug==3.1.3 ; python_full_version == "3.11.2"
yarl==1.20.1 ; python_full_version == "3.11.2"
